# Documentation/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/suppliers/{id}/balance": {
      "get": {
        "operationId": "SuppliersController_getBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Supplier UUID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierBalanceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Supplier not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get supplier balance",
        "tags": ["Suppliers"]
      }
    },
...
    "/api/v1/customers/{id}/balance": {
      "get": {
        "operationId": "CustomersController_getBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer UUID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerBalanceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get customer balance",
        "tags": ["Customers"]
      }
    },
...
    "/api/v1/products/{id}/stock": {
      "get": {
        "operationId": "ProductsController_getStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product UUID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Product stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductStockResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get product stock and average cost",
        "tags": ["Products"]
      }
    },
...
    "/api/v1/payment-accounts/{id}/balance": {
      "get": {
        "operationId": "PaymentAccountsController_getBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment account UUID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment account balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentAccountBalanceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Payment account not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get payment account balance",
        "tags": ["Payment Accounts"]
      }
    },
...
    "/api/v1/reports/suppliers/{id}/balance": {
      "get": {
        "operationId": "ReportsController_getSupplierBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Supplier UUID",
            "schema": { "type": "string" }
          },
          {
            "name": "asOfDate",
            "required": false,
            "in": "query",
            "description": "Point-in-time date (defaults to today)",
            "schema": { "example": "2026-02-20", "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Supplier balance with breakdown" },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Supplier not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Supplier balance report (point-in-time)",
        "tags": ["Reports"]
      }
    },
    "/api/v1/reports/customers/{id}/balance": {
      "get": {
        "operationId": "ReportsController_getCustomerBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer UUID",
            "schema": { "type": "string" }
          },
          {
            "name": "asOfDate",
            "required": false,
            "in": "query",
            "description": "Point-in-time date (defaults to today)",
            "schema": { "example": "2026-02-20", "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Customer balance with breakdown" },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Customer balance report (point-in-time)",
        "tags": ["Reports"]
      }
    },
    "/api/v1/reports/payment-accounts/{id}/balance": {
      "get": {
        "operationId": "ReportsController_getPaymentAccountBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment account UUID",
            "schema": { "type": "string" }
          },
          {
            "name": "asOfDate",
            "required": false,
            "in": "query",
            "description": "Point-in-time date (defaults to today)",
            "schema": { "example": "2026-02-20", "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Payment account balance with breakdown" },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Payment account not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Payment account balance report (point-in-time)",
        "tags": ["Reports"]
      }
    },
    "/api/v1/reports/products/{id}/stock": {
      "get": {
        "operationId": "ReportsController_getProductStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product UUID",
            "schema": { "type": "string" }
          },
          {
            "name": "asOfDate",
            "required": false,
            "in": "query",
            "description": "Point-in-time date (defaults to today)",
            "schema": { "example": "2026-02-20", "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Product stock with movement breakdown" },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Product stock report (point-in-time)",
        "tags": ["Reports"]
      }
    },
    "/api/v1/reports/pending-receivables": {
      "get": {
        "operationId": "ReportsController_getPendingReceivables",
        "parameters": [
          {
            "name": "asOfDate",
            "required": false,
            "in": "query",
            "description": "Point-in-time date (defaults to today)",
            "schema": { "example": "2026-02-20", "type": "string" }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "description": "Filter to a single customer",
            "schema": { "type": "string" }
          },
          {
            "name": "minAmount",
            "required": false,
            "in": "query",
            "description": "Minimum balance threshold (default 0)",
            "schema": { "example": 0, "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending receivables list with open documents"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Pending receivables — all customers with outstanding AR balances",
        "tags": ["Reports"]
      }
    },
    "/api/v1/reports/pending-payables": {
      "get": {
        "operationId": "ReportsController_getPendingPayables",
        "parameters": [
          {
            "name": "asOfDate",
            "required": false,
            "in": "query",
            "description": "Point-in-time date (defaults to today)",
            "schema": { "example": "2026-02-20", "type": "string" }
          },
          {
            "name": "supplierId",
            "required": false,
            "in": "query",
            "description": "Filter to a single supplier",
            "schema": { "type": "string" }
          },
          {
            "name": "minAmount",
            "required": false,
            "in": "query",
            "description": "Minimum balance threshold (default 0)",
            "schema": { "example": 0, "type": "number" }
          }
        ],
        "responses": {
          "200": { "description": "Pending payables list with open documents" },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Pending payables — all suppliers with outstanding AP balances",
        "tags": ["Reports"]
      }
    },
    "/api/v1/reports/suppliers/{id}/statement": {
      "get": {
        "operationId": "ReportsController_getSupplierStatement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Supplier UUID",
            "schema": { "type": "string" }
          },
          {
            "name": "dateFrom",
            "required": true,
            "in": "query",
            "description": "Start date inclusive (YYYY-MM-DD)",
            "schema": { "example": "2026-01-01", "type": "string" }
          },
          {
            "name": "dateTo",
            "required": true,
            "in": "query",
            "description": "End date inclusive (YYYY-MM-DD)",
            "schema": { "example": "2026-02-20", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier statement with opening/closing balance and ledger entries"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Supplier not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Supplier account statement with running balance",
        "tags": ["Reports"]
      }
    },
    "/api/v1/reports/customers/{id}/statement": {
      "get": {
        "operationId": "ReportsController_getCustomerStatement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer UUID",
            "schema": { "type": "string" }
          },
          {
            "name": "dateFrom",
            "required": true,
            "in": "query",
            "description": "Start date inclusive (YYYY-MM-DD)",
            "schema": { "example": "2026-01-01", "type": "string" }
          },
          {
            "name": "dateTo",
            "required": true,
            "in": "query",
            "description": "End date inclusive (YYYY-MM-DD)",
            "schema": { "example": "2026-02-20", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer statement with opening/closing balance and ledger entries"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Customer account statement with running balance",
        "tags": ["Reports"]
      }
    },
    "/api/v1/reports/payment-accounts/{id}/statement": {
      "get": {
        "operationId": "ReportsController_getPaymentAccountStatement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment account UUID",
            "schema": { "type": "string" }
          },
          {
            "name": "dateFrom",
            "required": true,
            "in": "query",
            "description": "Start date inclusive (YYYY-MM-DD)",
            "schema": { "example": "2026-01-01", "type": "string" }
          },
          {
            "name": "dateTo",
            "required": true,
            "in": "query",
            "description": "End date inclusive (YYYY-MM-DD)",
            "schema": { "example": "2026-02-20", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment account statement with opening/closing balance"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Payment account not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Payment account statement with running balance",
        "tags": ["Reports"]
      }
    },
...
  },
  "info": {
    "title": "Persona Finance API",
    "description": "Multi-tenant finance system. Standard error response format includes: `statusCode`, `message`, `errors[]`, `timestamp`, `path`, `requestId`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "ApiErrorField": {
        "type": "object",
        "properties": {
          "field": { "type": "string", "example": "name" },
          "message": { "type": "string", "example": "must not be empty" }
        },
        "required": ["field", "message"]
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 400 },
          "message": { "type": "string", "example": "Validation failed" },
          "errors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ApiErrorField" }
          },
          "timestamp": {
            "type": "string",
            "example": "2026-02-11T12:00:00.000Z"
          },
          "path": { "type": "string", "example": "/api/v1/customers" },
          "requestId": { "type": "string", "example": "req_123" }
        },
        "required": ["statusCode", "message", "timestamp", "path"]
      },
...
      "SupplierBalanceResponseDto": {
        "type": "object",
        "properties": {
          "supplierId": {
            "type": "string",
            "example": "9f4b6e2c-0a2d-4cc5-8c4d-1a4a88c81a88"
          },
          "totalPurchases": {
            "type": "number",
            "example": 150000,
            "description": "Total purchases in PKR (integer)"
          },
          "totalPaid": {
            "type": "number",
            "example": 50000,
            "description": "Total paid in PKR (integer)"
          },
          "currentBalance": {
            "type": "number",
            "example": 100000,
            "description": "Current balance in PKR (integer)"
          }
        },
        "required": [
          "supplierId",
          "totalPurchases",
          "totalPaid",
          "currentBalance"
        ]
      },
...
      "CustomerBalanceResponseDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "example": "9f4b6e2c-0a2d-4cc5-8c4d-1a4a88c81a88"
          },
          "totalSales": {
            "type": "number",
            "example": 150000,
            "description": "Total sales in PKR (integer)"
          },
          "totalPayments": {
            "type": "number",
            "example": 40000,
            "description": "Total payments received in PKR (integer)"
          },
          "totalReturns": {
            "type": "number",
            "example": 10000,
            "description": "Total return credits in PKR (integer)"
          },
          "currentBalance": {
            "type": "number",
            "example": 100000,
            "description": "Current balance in PKR (integer)"
          }
        },
        "required": [
          "customerId",
          "totalSales",
          "totalPayments",
          "totalReturns",
          "currentBalance"
        ]
      },
...
      "ProductStockResponseDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "9f4b6e2c-0a2d-4cc5-8c4d-1a4a88c81a88"
          },
          "productName": { "type": "string", "example": "Sella Rice 25kg" },
          "currentStock": { "type": "number", "example": 250 },
          "avgCost": {
            "type": "number",
            "example": 125000,
            "description": "Average cost in PKR (integer)"
          }
        },
        "required": ["productId", "productName", "currentStock", "avgCost"]
      },
...
      "PaymentAccountBalanceResponseDto": {
        "type": "object",
        "properties": {
          "paymentAccountId": {
            "type": "string",
            "example": "9f4b6e2c-0a2d-4cc5-8c4d-1a4a88c81a88"
          },
          "openingBalance": {
            "type": "number",
            "example": 25000,
            "description": "Opening balance in PKR (integer)"
          },
          "totalIn": {
            "type": "number",
            "example": 150000,
            "description": "Total in amount in PKR (integer)"
          },
          "totalOut": {
            "type": "number",
            "example": 50000,
            "description": "Total out amount in PKR (integer)"
          },
          "currentBalance": {
            "type": "number",
            "example": 125000,
            "description": "Current balance in PKR (integer)"
          }
        },
        "required": [
          "paymentAccountId",
          "openingBalance",
          "totalIn",
          "totalOut",
          "currentBalance"
        ]
      },
...
    }
  }
}